fix(tests): centralize firestore mock data
